component + apiReq refactoring
