+++ src/js/engagement_view/src/components/login/LogIn.css
@@ -3,7 +3,7 @@
 .backgroundImage{
   width: 100vw;
   height: 100vh;
-  background: url("./images/nodes.jpg");
+  background: url("../../images/nodes.jpg");
   background-position: center;
   background-repeat: no-repeat;
   background-size: cover;
