+++ src/js/engagement_view/src/components/login/styles.tsx
@@ -1,0 +1,14 @@
+import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
+
+export const loginStyles = makeStyles(
+    (theme: Theme) =>
+        createStyles(
+            {
+                valErrorMsg:{
+                    marginLeft: ".8rem",
+                    color: "red",
+                    fontSize: ".75rem"
+                }
+            }
+        )
+);
