+++ src/js/engagement_view/src/components/reusableComponents/graplHeader/styles.tsx
@@ -1,0 +1,18 @@
+import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
+
+export const graplHeaderStyles = makeStyles(
+    (theme: Theme) =>
+        createStyles({
+            root: {
+                display: "flex",
+            },
+            header: {
+                justifyContent: "space-between",
+            }, 
+            link:{
+                color:"#42C6FF", 
+                textDecoration: "none"
+            }
+        }
+    )
+);
