+++ src/js/engagement_view/src/components/reusableComponents/graplHeader/types.tsx
@@ -1,0 +1,3 @@
+export type GraplHeaderProps = {
+    displayBtn: boolean
+}
