+++ src/js/engagement_view/src/components/reusableComponents/index.ts
@@ -1,4 +1,4 @@
-import GraplHeader from "./GraplHeader";
-import LoginNotification from "./Notifications";
+import GraplHeader from "./graplHeader/GraplHeader";
+import LoginNotification from "./notifications/Notifications";
 
 export { GraplHeader, LoginNotification };
