+++ src/js/engagement_view/src/components/reusableComponents/notifications/styles.tsx
@@ -1,0 +1,12 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const notificationsStyles = makeStyles({
+    root: {
+        maxWidth: 345,
+        postion: "fixed",
+    },
+    button: {
+        border: "2px solid white",
+        backgroundColor: "#42c6ff",
+    }
+});
