+++ src/js/engagement_view/src/components/uploadPlugin/uploadPluginStyles.tsx
